% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{brmsmargins}
\alias{brmsmargins}
\title{brmsmargins}
\usage{
brmsmargins(
  m,
  cond = expand.grid(Time = 1, Tx = 0:1),
  ci,
  CIType,
  prob = TRUE,
  rope,
  mid
)
}
\arguments{
\item{m}{model}

\item{cond}{stuff}

\item{ci}{ci}

\item{CIType}{the type}

\item{prob}{is it a probability}

\item{rope}{rope}

\item{mid}{mid}
}
\description{
brmsmargins
}
