% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check.brmsfit}
\alias{check.brmsfit}
\title{Function to check that something is a brmsfit object}
\usage{
check.brmsfit(object)
}
\arguments{
\item{object}{An object to be evaluated.}
}
\value{
An informative error if not of class \code{brmsfit},
otherwise invisibly returns \code{TRUE}.
}
\description{
Internal utility function to check, with a useful error if not,
whether and object has class \code{brmsfit}.
}
\keyword{internal}
