% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check.gaussian}
\alias{check.gaussian}
\title{Function to check whether a brmsfit with random effects are all gaussian}
\usage{
check.gaussian(object)
}
\arguments{
\item{object}{An object to be evaluated.}
}
\value{
\code{TRUE} if all random effects present are Gaussian or
if there are no random effects.
\code{FALSE} if any random effect present is not Gaussian.
}
\description{
Internal utility function to check whether a \code{brmsfit}
object has all gaussian random effects.
}
\keyword{internal}
