% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.extractlink}
\alias{.extractlink}
\title{Extract the Link from a \code{brms} Model}
\usage{
.extractlink(object, dpar)
}
\arguments{
\item{object}{A \link[brms:brmsfit-class]{brms::brmsfit} model object.}

\item{dpar}{The dpar for which the link should be extracted.}
}
\value{
A character string, the link.
}
\description{
Internal utility function to take a \link[brms:brmsfit-class]{brms::brmsfit} object
and extract the link for a specific brms \code{dpar}.
}
\keyword{internal}
