% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.links}
\alias{.links}
\title{Internal function to build out scale, inverselink and function}
\usage{
.links(
  link,
  effects = c("fixedonly", "includeRE", "integrateoutRE"),
  backtrans = c("response", "linear", "identity", "invlogit", "exp", "square")
)
}
\arguments{
\item{link}{The link named in a \code{brmsfit} object}

\item{effects}{A character string, the type of effect desired}

\item{backtrans}{A character string, the type of back transformation}
}
\value{
A list. TODO.
}
\description{
Internal utility function used in \code{.predict()}
}
\keyword{internal}
