% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.lnames}
\alias{.lnames}
\title{Function to generate names of an L matrix from brms}
\usage{
.lnames(block, ncol)
}
\arguments{
\item{block}{An integer, which random effect block object to be used.}

\item{ncol}{An integer, the total number of columns. Since this is a square matrix,
this also defines how many rows.}
}
\value{
A character string of the names.
}
\description{
Internal utility function to check create the variable names for the
Cholesky decomposition of the random effects correlation matrix in \code{brms}.
\code{brms} returns the lower triangular matrix and we want the upper triangular
matrix, so the names are transposed. The results can then be passed to
the \code{tab2mat} function to convert the row vector into a matrix.
}
\keyword{internal}
