% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.predict}
\alias{.predict}
\title{Function to generate posterior predictions and summaries from a brms model}
\usage{
.predict(
  object,
  data,
  summarize = TRUE,
  posterior = FALSE,
  dpar = NULL,
  re_formula = NULL,
  resample = 0L,
  seed,
  ...
)
}
\arguments{
\item{object}{A fitted brms model object. Required.}

\item{data}{A data frame or data table passed to \code{fitted()}
as the new data to be used for predictions. Required.}

\item{summarize}{A logical value, whether or not to
calculate summaries of the posterior predictions.
Defaults to \code{TRUE}.}

\item{posterior}{A logical value whether or not to
save and return the posterior samples. Defaults
to \code{FALSE} as the assumption is a typical
use case is to return the summaries only.}

\item{dpar}{Parameter passed on the \code{dpar}
argument of \code{fitted()} in brms.}

\item{re_formula}{Parameter passed on the \code{re_formula}
argument of \code{fitted()} in brms.}

\item{resample}{An integer indicating the number of
bootstrap resamples of the posterior predictions to
use when calculating summaries. Defaults to \code{0L}.
See the details section for more informations as its implementation
is experimental and it may not operate as one would expect.}

\item{seed}{A seed for random number generation. Missing by default.
Only needed if \code{resample} is a non zero integer.}

\item{...}{Additional arguments passed to \code{fitted()}}
}
\value{
A list with \code{Summary} and \code{Posterior}.
Some of these may be \code{NULL} depending on the arguments used.
}
\description{
This is an internal function that is essentially a fancy wrapper for
\code{fitted()}.
}
\keyword{internal}
