% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{integratere}
\alias{integratere}
\title{To write}
\usage{
integratere(obj)
}
\arguments{
\item{obj}{A list}
}
\value{
A numeric matrix with random values
}
\description{
TODO: write description.
}
\examples{

}
