% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.random}
\alias{is.random}
\title{Check whether a \code{brmsfit} class object has random effects}
\usage{
is.random(object)
}
\arguments{
\item{object}{An object to be evaluated.}
}
\value{
\code{TRUE} if any random effects present.
\code{FALSE} if no random effects present.
}
\description{
Internal utility function to check whether a \code{brmsfit}
object has any random effects or not.
}
\keyword{internal}
