% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lmcpp_alt}
\alias{lmcpp_alt}
\title{Fast(er?) Linear Regression}
\usage{
lmcpp_alt(X, y)
}
\arguments{
\item{X}{A numeric model matrix. If intercept is desired, it must already have been added as a column.}

\item{y}{A numeric matrix. A single column if one response variable or multiple columns
where each column is a different response, such as a for marginal coefficients where
each column is a different MCMC sample.}
}
\value{
A numeric matrix with the coefficient.
}
\description{
Used to get marginal coefficients off of a generalized linear mixed model.
}
\examples{
lmcpp_alt(cbind(1, mtcars$hp, mtcars$am), as.matrix(mtcars[, c("mpg", "qsec")]))
}
