% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tab2mat}
\alias{tab2mat}
\title{Convert a row of a table to a matrix}
\usage{
tab2mat(X)
}
\arguments{
\item{X}{a matrix}
}
\value{
A numeric matrix with one row.
}
\description{
TODO: write description.
}
\examples{

}
