% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{builders}
\alias{builders}
\alias{.namesL}
\alias{.buildL}
\alias{.namesSD}
\alias{.buildSD}
\alias{.namesZ}
\alias{.buildZ}
\title{Internal functions build the variable names or data objects for estimation}
\usage{
.namesL(block, number)

.buildL(data, block, number, dpar)

.namesSD(ranef, block, dpar)

.buildSD(data, ranef, block, dpar)

.namesZ(block, number, dpar)

.buildZ(data, block, number, dpar)
}
\arguments{
\item{block}{Which random effect block to use. An integer.}

\item{number}{The number of elements in that random effect block. An integer.}

\item{data}{A data object. For example the result of \code{\link[=make_standata]{make_standata()}}
for \code{\link[=.buildZ]{.buildZ()}}, which is a list,
or a dataset of the posterior draws such as from \code{\link[=as_draws_df]{as_draws_df()}}
for \code{\link[=.buildL]{.buildL()}} and \code{\link[=.buildSD]{.buildSD()}}.}

\item{dpar}{Which dpar to use. Does not apply to the L matrix.}

\item{ranef}{A data set with information about the model object random effects.
Only used for \code{.namesSD} and \code{.buildSD}.}
}
\value{
A character vector for all \code{.names} functions or a matrix
for all \code{.build} functions.
}
\description{
These are a set of internal utility functions.
They are not intended for general use.
}
\details{
\itemize{
\item{\code{.namesL}}{Generate names of an L matrix from \code{brms}. Create the variable names for the Cholesky decomposition of the random effects correlation matrix in \code{brms}. Note that \code{brms} returns the lower triangular matrix and we want the upper triangular matrix, so the names are transposed. The results can then be passed to the \code{tab2mat} function to convert the row vector into a matrix.}
\item{\code{.buildL}}{Returns the L matrix object. Rows are posterior draws.}
\item{\code{.namesSD}}{Create the names of random effect standard deviation estimates.}
\item{\code{.buildSD}}{Return matrix of random effect standard deviation estimates. Rows are posterior draws.}
\item{\code{.namesZ}}{Create the names of random effects data for predictions.}
\item{\code{.buildZ}}{Return matrix of data for random effect predictions.}
}
}
\examples{
brmsmargins:::.namesL(1, 3)
brmsmargins:::tab2matR(matrix(brmsmargins:::.namesL(1, 3), 1))
brmsmargins:::.namesZ(1, 3, NULL)
}
\keyword{internal}
