% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checktab}
\alias{.checktab}
\title{Check Object Class is a Table}
\usage{
.checktab(x, requireNames = TRUE)
}
\arguments{
\item{x}{An object to be evaluated.}

\item{requireNames}{A logical, whether names are
required. Defaults to \code{TRUE}}
}
\value{
An empty string if no issues. Otherwise, a non zero
string with warning/error messages.
}
\description{
Internal utility function confirm that an object
has the attributes needed to be used as data.
Currently it should be a \code{tbl},
\code{data.frame}, or \code{data.table}.
}
\keyword{internal}
