% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checktab}
\alias{.checktab}
\title{Function to check that something is a valid data object}
\usage{
.checktab(x, requireNames = TRUE)
}
\arguments{
\item{x}{An object to be evaluated.}

\item{requireNames}{A logical, whether names are
required. Defaults to \code{TRUE}}
}
\value{
An empty string if no issues. Otherwise, a non zero
string with warning/error messages.
}
\description{
Internal utility function to check requirements for
\code{\link{brmsmargins}}.
}
\keyword{internal}
