% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{integratere}
\alias{integratere}
\title{Integrate over Random Effects}
\usage{
integratere(d, sd, L, k, yhat)
}
\arguments{
\item{d}{A list}

\item{sd}{A list}

\item{L}{A list}

\item{k}{An integer, the number of samples}

\item{yhat}{A matrix of the fixed effects predictions}
}
\value{
A numeric matrix with random values
}
\description{
TODO: write description.
}
\examples{

}
