% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{integratere}
\alias{integratere}
\title{Integrate over Random Effects}
\usage{
integratere(d, sd, L, k, yhat, backtrans)
}
\arguments{
\item{d}{A list}

\item{sd}{A list}

\item{L}{A list}

\item{k}{An integer, the number of samples}

\item{yhat}{A matrix of the fixed effects predictions}

\item{backtrans}{An integer, indicating the type of back transformation.
0 indicates inverse logit (e.g., for logistic regression).
1 indicates exponential (e.g., for poisson or negative binomial regression or if outcome was natural log transformed).
2 indicates square (e.g., if outcome was square root transformed).}
}
\value{
A numeric matrix with random values
}
\description{
TODO: write description.
}
\examples{

}
