% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tab2mat}
\alias{tab2mat}
\title{To write}
\usage{
tab2mat(X, index)
}
\arguments{
\item{obj}{A list}
}
\value{
A numeric matrix with random values
}
\description{
TODO: write description.
}
\examples{

}
