% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rimplementation.R
\name{tab2mat}
\alias{tab2mat}
\alias{tab2matR}
\title{Convert a row of a table to a matrix}
\usage{
tab2mat(X)

tab2matR(X)
}
\arguments{
\item{X}{a matrix}
}
\value{
A numeric matrix with one row.
}
\description{
TODO: write description.
}
\section{Functions}{
\itemize{
\item \code{tab2matR}: Pure \code{R} implementation of \code{tab2mat}
}}

\examples{

}
