% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rimplementation.R
\name{tab2mat}
\alias{tab2mat}
\alias{tab2matR}
\title{Convert a Row of a Table to a Square Matrix}
\usage{
tab2mat(X)

tab2matR(X)
}
\arguments{
\item{X}{a matrix}
}
\value{
A numeric matrix with one row.
}
\description{
Utility function to convert a row matrix to a square matrix.
Used as the \code{brms} package returns things like the Cholesky
decomposition matrix as separate columns where rows are posterior draws.
Not intended to be called directly by most users.
}
\section{Functions}{
\itemize{
\item \code{tab2matR()}: Pure \code{R} implementation of \code{tab2mat}

}}
\examples{
tab2mat(matrix(1:4, 1))
tab2mat(matrix(1:9, 1))
}
